- rename to brewpi_webservice - rename package `rest` to `auth' - merge auth urls
