+++ brewpi_webservice/admin.py
@@ -4,6 +4,9 @@ from django.contrib.auth.admin import UserAdmin, GroupAdmin
 from django.utils.translation import ugettext as _
 
 class BrewPiAdminSite(AdminSite):
+    """
+    Custom Admin site for BrewPi
+    """
     site_header = _("BrewPi Webservice Administration")
 
 admin_site = BrewPiAdminSite(name='brewpi-admin')
