+++ brewpi_webservice/authentication/views.py
@@ -1,6 +1,8 @@
 from django.contrib.auth.models import User, Group
+
 from rest_framework import viewsets
-from brewpiweb.rest.serializers import UserSerializer, GroupSerializer
+
+from .serializers import UserSerializer, GroupSerializer
 
 
 class UserViewSet(viewsets.ModelViewSet):
