+++ brewpi_webservice/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for brewpiweb project.
+Django settings for brewpi_webservice project.
 
 Generated by 'django-admin startproject' using Django 1.9.2.
 
@@ -37,9 +37,12 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    # Third-party libraries
     'django_extensions',
     'rest_framework',
     'rest_framework_swagger',
+    # Local apps
+    'brewpi_webservice.authentication',
 ]
 
 MIDDLEWARE_CLASSES = [
@@ -53,7 +56,7 @@ MIDDLEWARE_CLASSES = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'brewpiweb.urls'
+ROOT_URLCONF = 'brewpi_webservice.urls'
 
 TEMPLATES = [
     {
@@ -71,7 +74,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'brewpiweb.wsgi.application'
+WSGI_APPLICATION = 'brewpi_webservice.wsgi.application'
 
 
 # Database
