+++ brewpi_webservice/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for brewpiweb project.
+WSGI config for brewpi_webservice project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "brewpiweb.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "brewpi_webservice.settings")
 
 application = get_wsgi_application()
